+++ python/src/multigroup/EnergyGroupStructure.python.cpp
